feat: 🎸 use compute+execute split instead of compute only
