+++ server/src/internal/billing/v2/execute/executeAutumnBillingPlan.ts
@@ -1,7 +1,7 @@
 import type { AutumnBillingPlan, Invoice } from "@autumn/shared";
 import type Stripe from "stripe";
 import type { AutumnContext } from "@/honoUtils/HonoEnv";
-import { applyAutoTopupRebalance } from "@/internal/billing/v2/execute/executeAutumnActions/applyAutoTopupRebalance";
+import { executeAutoTopupRebalance } from "@/internal/billing/v2/execute/executeAutumnActions/executeAutoTopupRebalance";
 import { insertNewCusProducts } from "@/internal/billing/v2/execute/executeAutumnActions/insertNewCusProducts";
 import {
 	getDeleteCustomerProducts,
@@ -107,14 +107,13 @@ export const executeAutumnBillingPlan = async ({
 		updates: autumnBillingPlan.updateCustomerEntitlements,
 	});
 
-	// 5a. Auto top-up rebalance (post-Stripe, live-read paydown).
-	// Reads live cusEnt balances and applies atomic delta increments, so concurrent
-	// usage between the ATU context snapshot and this write is preserved.
+	// 5a. Auto top-up rebalance: apply pre-computed paydown + remainder deltas as
+	// atomic SQL `balance + delta` increments.
 	if (autumnBillingPlan.autoTopupRebalance) {
-		await applyAutoTopupRebalance({
+		await executeAutoTopupRebalance({
 			ctx,
 			customerId: autumnBillingPlan.customerId,
-			intent: autumnBillingPlan.autoTopupRebalance,
+			deltas: autumnBillingPlan.autoTopupRebalance.deltas,
 		});
 	}
