+++ shared/models/billingModels/plan/autumnBillingPlan.ts
@@ -86,25 +86,18 @@ export const AutumnBillingPlanSchema = z.object({
 		.optional(),
 
 	/**
-	 * Declarative auto top-up rebalance intent.
-	 *
-	 * When present, the executor (post-Stripe-charge) reads LIVE cusEnt balances for
-	 * `featureId`, pays down any existing overage first (bounded by `quantity` and the
-	 * depth of overage), and routes the remainder to the prepaid cusEnt identified by
-	 * `prepaidCustomerEntitlementId`.
-	 *
-	 * Live-read avoids two races inherent to snapshot-based paydown:
-	 *   - Concurrent usage between snapshot and execution being overwritten (data loss).
-	 *   - Concurrent refunds making the snapshot's overage look deeper than it actually is
-	 *     (would produce overcredit).
-	 *
-	 * All writes are atomic SQL `balance + delta` increments via adjustBalanceDbAndCache.
+	 * Pre-computed auto top-up rebalance deltas. The compute step sizes paydown + prepaid
+	 * remainder from the context's FullCustomer snapshot; the executor just loops these
+	 * and applies each via adjustBalanceDbAndCache (atomic SQL balance + delta).
 	 */
 	autoTopupRebalance: z
 		.object({
-			featureId: z.string(),
-			quantity: z.number(),
-			prepaidCustomerEntitlementId: z.string(),
+			deltas: z.array(
+				z.object({
+					cusEntId: z.string(),
+					delta: z.number(),
+				}),
+			),
 		})
 		.optional(),
